requirements.txt
openai
langchain
langchain-openai
langchain-core
langchain-community
langchain-groq
langchainhub

streamlit
python-dotenv
requests


faiss-cpu
chromadb

langserve
transformers

yt_dlp